instructions on dash
cant crouch when dashing, doesnt do effects of dashing during the cooldown. cooldown has to finish for the effects to happen. however this causes the effect to not work if the player has started crouching before the cooldown has reached zero. causing the effect to not happen unless crouched after cooldown resets. small gripe but enough for annoyance.